fixes #101 fall through next middleware if route not found
Just in case someone wants to add some extra middleware after this one.
Express will take care of issuing the 404 Not found message.

Also, deals with redirection as per router documentation.
https://github.com/reactjs/react-router-redux/blob/master/examples/server/server.js#L46

Also, if I don't know what a callback would do with the return, I prefer to do `return void next()`.  In this case, though, it is safe since react-router doesn't care, but I had to check it out.  Simpler to add the `void` and be safe.